Qt-KDE Wrapper (http://iidea.pl/~js/qkw/)  code (KDElibs for win32) merged with current kdeui/
Notes:
-K_WS_QTONLY macro removed, some ifdefs removed
-some bugs fixed


-kaboutapplication.cpp     -logo is set using KAboutData::programLogo()
-kaboutdialog.cpp       -added setIcon( const QPixmap& )
-kiconview.cpp   -see klistbox.cpp
-klistbox.cpp   -some thing wont't work because currently
                  KApplication::keyboardModifiers() always returns 0
-klistview.cpp   -see klistbox.cpp
-kmainwindow.cpp    -QWidet::isMaximized() is used instead of NET
-kmessagebox.cpp    -only ifdefed XSetTransientForHint()
-kpixmapio.cpp  -no shared memory used
-krootpixmap.cpp   -compiles, not working (lack of KSharedPixmap)
-ksharedpixmap.cpp    -won't work unless pixmap server is implemented
-kwindowinfo.cpp   -only partially working due to lack of KWin
-kwindowlistmenu.cpp   -x11 only, can be implemented later
-qxembed.cpp    -only used on X11

svn path=/trunk/kdelibs/; revision=345070

+++ kdeui/kcombobox.cpp
@@ -556,7 +556,7 @@ void KHistoryCombo::rotateDown()
         }
         else { // bottom of history
             if ( myIterateIndex == -2 ) {
-                KNotifyClient::event( winId(), KNotifyClient::notification,
+                KNotifyClient::event( (int)winId(), KNotifyClient::notification,
                                       i18n("No further item in the history."));
             }
